refactor: clean up kvk history command flow
